% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcon.R
\name{pcon}
\alias{pcon}
\title{Planned contrast analysis}
\usage{
pcon(dat, w, iv_n = 0)
}
\arguments{
\item{dat}{A matrix object. In the case of a between-persons design it needs
to include a single between-persons variable in the first column and a
dependent variable in the second column. In the case of a}

\item{w}{A numeric vector object including the contrast weights}

\item{iv_n}{The number of between-persons variables in data. iv_n can be either 0 or 1.}
}
\value{
A list object including all information association with the planned contrast tests.
}
\description{
The function performs planned contrast analysis using the formulas provides
in Maxwell and Delaney (2004) and Rosenthal and Rosnow (1985).
}
\examples{
#Example 1 (between-persons design with unequal cell n)

cond <- as.factor(c(1, 1, 1, 1, 2, 2, 2, 2, 3, 3, 3, 3, 4, 4, 4, 4, 4))
score <- c(3, 5, 1, 7, 3, 2, 2, 9, 4, 5, 7, 4, 9, 4, 1, 8, 5)

pcon(cbind(cond, score), w = c(-3, -1, 1, 3), iv_n = 1)

}
\references{
Rosenthal, R., & Rosnow, R. L. (1981). \emph{Contrast analysis: Focused comparisons
in the analysis of variance}. Cambridge University Press.

Maxwell, S. E., & Delaney, H. D. (2004). \emph{Designing experiments and
analyzing data. A model comparison perspective} (2nd ed.). Lawrence Erlbaum.
}
